requirements.txt
Flask
rdkit-pypi
scikit-learn
joblib
numpy
pandas
plotly